Cleanup codebase
