+++ pages/index.js
@@ -1,616 +1,6 @@
 import React, { useState } from 'react'
-import styled from 'styled-components'
-import axios from 'axios'
-import { space, layout, typography, border, color } from 'styled-system'
-import {
-  Box,
-  Menu,
-  MenuItem,
-  Title,
-  Text,
-  IconGlif
-} from '../components/Shared'
-
-const TitleCopy = styled(Title)`
-  /* Used this: https://stackoverflow.com/questions/14431411/pure-css-to-make-font-size-responsive-based-on-dynamic-amount-of-characters */
-  font-size: calc(48px + (64 - 40) * (100vw - 360px) / (1440 - 360));
-`
-
-const Image = styled.img`
-  z-index: -999;
-`
-
-const ButtonSignUp = styled.button`
-  outline: none;
-  border: 0;
-  cursor: pointer;
-  background: transparent;
-  transition: 0.18s ease-in;
-
-  &:hover {
-    transform:scale(1.05);
-  }
-  ${space}
-  ${layout}
-  ${typography}
-  ${border} 
-  ${color}
-`
-
-const InputEmail = styled.input`
-  outline: 0;
-  border: 0;
-  ::placeholder {
-    color: #444;
-  }
-  ${space}
-  ${layout}
-  ${typography}
-  ${border}
-  ${color}
-`
-
-const ExtLink = styled.a`
-  position: relative;
-  text-decoration: none;
-  color: #0051ff;
-  background: #0051ff00;
-  z-index: 9;
-  transition: 0.24s ease-in-out;
-  ${space}
-  ${layout}
-  ${typography}
-  ${border}
-  ${color}
-
-  &:hover {
-    background: #0051ff;
-    color: #fff;
-    
-  }
-
-  &:before {
-    content: '';
-    background-size: contain;
-    background-repeat: no-repeat
-    background-position: center center;
-    border: 1px solid;
-
-    opacity: 0.5;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    border-radius: 16px;
-    position: absolute;
-    z-index: -9;
-  }
-
-
-`
+import Home from '../components/Home'
 
 export default () => {
-  const [clicked, setClicked] = useState(false)
-  const [error, setError] = useState('')
-  const [subscribed, setSubscribed] = useState(false)
-  const [email, setEmail] = useState('')
-  const postToMailChimp = async () => {
-    try {
-      const res = await axios.post(
-        `https://mailchimp-proxy.openworklabs.com/${email}`
-      )
-
-      if (res.data.indexOf('success') === -1) {
-        if (res.data.indexOf('already subscribed') > -1) {
-          setError("You're already subscribed. :)")
-          return
-        }
-
-        setError(
-          "There was an issue getting you subscribed. We're on the case!"
-        )
-      } else {
-        setSubscribed(true)
-        setError('')
-      }
-    } catch (error) {
-      setError(error.toString())
-    }
-  }
-
-  return (
-    <Box
-      display='block'
-      margin='0 auto'
-      py={[2, 4]}
-      pl={[3, 4, 6, 7]}
-      pr={[3, 4, 6]}
-      maxWidth='1680px'
-      width='100%'
-      minHeight='100vh'
-    >
-      <section name='Introduction'>
-        <Menu
-          display='flex'
-          alignItems='center'
-          justifyContent='flex-start'
-          flexWrap='wrap'
-        >
-          <MenuItem>
-            <Box
-              display='inline-block'
-              py={3}
-              mr={[4, 6, 7, 8]}
-              px={2}
-              borderRadius={4}
-              css={`
-                background: linear-gradient(180deg, #ff8d3b 0%, #ffeb80 100%);
-                border-radius: 16px;
-              `}
-            >
-              <IconGlif size={[6, 7, 8]} />
-            </Box>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>is</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>an</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>interoperable</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>set</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>of</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>tools</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>for</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>the</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>Filecoin</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>network.</TitleCopy>
-          </MenuItem>
-
-          <MenuItem
-            display='flex'
-            flexWrap='wrap'
-            alignItems='baseline'
-            justifyContent='space-between'
-            width='100%'
-            color='core.darkgray'
-            my={[2, 3]}
-          >
-            {clicked ? (
-              <>
-                <Menu display='flex' alignItems='center'>
-                  <MenuItem>
-                    <Box
-                      display='flex'
-                      flexWrap='wrap'
-                      width={['100%', 'auto']}
-                    >
-                      <InputEmail
-                        width={['100%', 'auto']}
-                        fontSize={[4, 5, 6]}
-                        color='core.nearblack'
-                        border={1}
-                        borderWidth={2}
-                        px={3}
-                        py={3}
-                        textAlign='center'
-                        placeholder='Your email, please'
-                        borderTopLeftRadius={[0, 2]}
-                        borderBottomLeftRadius={[0, 2]}
-                        onChange={(e) => setEmail(e.target.value)}
-                      />
-                      <ButtonSignUp
-                        width={['100%', 'auto']}
-                        color='core.white'
-                        bg='core.nearblack'
-                        fontSize={[4, 5, 6]}
-                        border={1}
-                        borderColor='core.nearblack'
-                        borderWidth={2}
-                        borderTopRightRadius={[0, 2]}
-                        borderBottomRightRadius={[0, 2]}
-                        px={6}
-                        py={3}
-                        height='max-content'
-                        onClick={postToMailChimp}
-                      >
-                        Submit
-                      </ButtonSignUp>
-                    </Box>
-
-                    {error ? (
-                      <Title mt={2} color='red'>
-                        {error}
-                      </Title>
-                    ) : (
-                      <Title mt={2}>
-                        Glif don&rsquo;t spam! Unsub whenever.
-                      </Title>
-                    )}
-                    {subscribed && (
-                      <Title mt={2} color='status.success.background'>
-                        You&rsquo;re subscribed. Keep an eye out.
-                      </Title>
-                    )}
-                  </MenuItem>
-                </Menu>
-                <ButtonSignUp
-                  width={['100%', 'auto']}
-                  background='transparent'
-                  color='core.nearblack'
-                  fontSize={[4, 5, 6]}
-                  border={1}
-                  px={6}
-                  py={2}
-                  my={2}
-                  height='max-content'
-                  borderRadius={6}
-                  onClick={() => setClicked(false)}
-                >
-                  Cancel
-                </ButtonSignUp>
-              </>
-            ) : (
-              <>
-                <Text fontSize={[4, 5, 6]} my={2}>
-                  Be the first to learn when we launch
-                </Text>
-
-                <ButtonSignUp
-                  background='transparent'
-                  color='core.nearblack'
-                  fontSize={[4, 5, 6]}
-                  border={1}
-                  px={6}
-                  py={2}
-                  height='max-content'
-                  borderRadius={4}
-                  onClick={() => setClicked(true)}
-                >
-                  Sign Up
-                </ButtonSignUp>
-              </>
-            )}
-          </MenuItem>
-        </Menu>
-      </section>
-      <section name='Glif Wallet'>
-        <Menu
-          display='flex'
-          alignItems='center'
-          justifyContent='flex-start'
-          flexWrap='wrap'
-          my={9}
-        >
-          <MenuItem display='flex' alignItems='center' width={['100%', 'auto']}>
-            <Box
-              display='flex'
-              alignItems='center'
-              justifyContent='space-between'
-              py={2}
-              mr={[0, 6]}
-              mb={[3, 0]}
-              px={3}
-              borderRadius={4}
-              width={['100%', 'auto']}
-              height={9}
-              css={`
-                background: url('/imgtools.png ') center center no-repeat;
-                background-size: 100%;
-                border-radius: 16px;
-              `}
-            >
-              <IconGlif
-                fill='#fff'
-                size={[6, 7, 8]}
-                css={`
-                  transform: rotate(-90deg);
-                `}
-              />
-              <TitleCopy ml={3}>Wallet</TitleCopy>
-            </Box>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>A</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>lightweight</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>interface</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>for</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>sending</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>and</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>receiving</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>Filecoin</TitleCopy>
-          </MenuItem>
-
-          {/* <MenuItem mr='8%' my={[2, 3]} height='120px'>
-            <Image width='300px' alt='' src='/imgwallet.png' />
-          </MenuItem> */}
-
-          <MenuItem width={['100%', 'auto']} mt={3}>
-            <ExtLink
-              href='https://wallet.glif.io'
-              display='flex'
-              borderRadius={4}
-              p={4}
-              fontSize={5}
-              css={`
-                /* Needs to use theme property */
-                border-radius: 16px;
-                ::before {
-                  background: url('/imgtools.png ');
-                  background-size: cover;
-                }
-              `}
-            >
-              <TitleCopy z-index='999'>
-                Go <sup>↗</sup>
-              </TitleCopy>
-            </ExtLink>
-          </MenuItem>
-        </Menu>
-      </section>
-      <section name='Glif Verify'>
-        <Menu
-          display='flex'
-          alignItems='center'
-          justifyContent='flex-start'
-          flexWrap='wrap'
-          my={9}
-        >
-          <MenuItem display='flex' alignItems='center' width={['100%', 'auto']}>
-            <Box
-              display='flex'
-              alignItems='center'
-              justifyContent='space-between'
-              py={2}
-              mr={[0, 4]}
-              mb={[3, 0]}
-              px={3}
-              borderRadius={4}
-              width={['100%', 'auto']}
-              height={9}
-              css={`
-                background: url('/imgverify.png') center no-repeat;
-                background-size: 110%;
-                border-radius: 16px;
-                alt: 'Source: https://www.nontemporary.com/post/190437968500';
-              `}
-            >
-              <IconGlif
-                fill='#fff'
-                size={[6, 7, 8]}
-                css={`
-                  transform: rotate(-90deg);
-                `}
-              />
-              <TitleCopy ml={3} color='white'>
-                Verify
-              </TitleCopy>
-            </Box>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>Earn</TitleCopy>
-          </MenuItem>
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>verified</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>Filecoin</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>storage</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>when</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>you</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>verify</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>yourself</TitleCopy>
-          </MenuItem>
-          <MenuItem width={['100%', 'auto']} mt={3}>
-            <ExtLink
-              href='https://verify.glif.io'
-              display='flex'
-              borderRadius={4}
-              p={4}
-              fontSize={5}
-              css={`
-                /* Needs to use theme property */
-                border-radius: 16px;
-                ::before {
-                  background: url('/imgverify.png ');
-                  background-size: 110%;
-                  background-position: center center;
-                }
-              `}
-            >
-              <TitleCopy z-index='999'>
-                Go <sup>↗</sup>
-              </TitleCopy>
-            </ExtLink>
-          </MenuItem>
-        </Menu>
-      </section>
-      <section name='Glif Node'>
-        <Menu
-          display='flex'
-          alignItems='center'
-          justifyContent='flex-start'
-          flexWrap='wrap'
-          my={9}
-        >
-          <MenuItem display='flex' alignItems='center' width={['100%', 'auto']}>
-            <Box
-              display='flex'
-              alignItems='center'
-              justifyContent='space-between'
-              py={2}
-              mr={[0, 4]}
-              mb={[3, 0]}
-              px={3}
-              borderRadius={4}
-              width={['100%', 'auto']}
-              height={9}
-              css={`
-                background: url('/imgnode.png') left center no-repeat;
-                border-radius: 16px;
-              `}
-              alt='Credit & Source: https://www.nontemporary.com/post/187451107349/rob-nick-carter'
-            >
-              <IconGlif
-                fill='#fff'
-                size={[6, 7, 8]}
-                css={`
-                  transform: rotate(-90deg);
-                `}
-              />
-              <TitleCopy ml={3}>Nodes</TitleCopy>
-            </Box>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>Public</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>and</TitleCopy>
-          </MenuItem>
-
-          {/* <MenuItem mr='8%' my={[2, 3]} height='120px'>
-            <Image
-              width='200px'
-              alt='Credit & Source: https://www.nontemporary.com/post/187451107349/rob-nick-carter'
-              src='/imgnode.png'
-            />
-          </MenuItem> */}
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy> dedicated</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>Filecoin</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy>node</TitleCopy>
-          </MenuItem>
-
-          <MenuItem mr='8%' my={[2, 3, 5]}>
-            <TitleCopy> infrastructure</TitleCopy>
-          </MenuItem>
-
-          <MenuItem width={['100%', 'auto']} mt={3}>
-            <ExtLink
-              href='https://wallet.glif.io'
-              display='flex'
-              borderRadius={4}
-              p={4}
-              fontSize={5}
-              css={`
-                /* Needs to use theme property */
-                border-radius: 16px;
-                ::before {
-                  background: url('/imgnode.png ');
-                  background-size: cover;
-                }
-              `}
-            >
-              <TitleCopy z-index='999'>
-                Go <sup>↗</sup>
-              </TitleCopy>
-            </ExtLink>
-          </MenuItem>
-        </Menu>
-      </section>
-      <section name='Made by OWL'>
-        <Menu>
-          <MenuItem display='flex' alignItems='center' my={[2, 3, 5]}>
-            <IconGlif size={6} />
-            <Title my={0} mx={2}>
-              is an
-              <ExtLink
-                href='https://www.openworklabs.com'
-                borderBottom={1}
-                borderWidth={3}
-                fontSize={4}
-                mx={1}
-                target='_blank'
-                css={`
-                  &:hover {
-                    color: #0051ff;
-                    background: transparent;
-                  }
-                  ::before {
-                    border: 0;
-                  }
-                `}
-              >
-                OWL
-              </ExtLink>{' '}
-              project {'\u00A9'} 2020
-            </Title>
-          </MenuItem>
-        </Menu>
-      </section>
-    </Box>
-  )
+  return <Home />
 }
